ut: add more lsp unit tests
